+++ test/index.test.ts
@@ -1,50 +1,13 @@
 import { expect, it, describe } from "vitest";
-import * as stdEnv from "../src";
+import * as stdEnv from "../src/index.ts";
 
 describe("std-env", () => {
-  it("has expected exports", () => {
-    expect(Object.keys(stdEnv)).toMatchInlineSnapshot(`
-      [
-        "env",
-        "nodeENV",
-        "platform",
-        "isCI",
-        "hasTTY",
-        "hasWindow",
-        "isDebug",
-        "isTest",
-        "isProduction",
-        "isDevelopment",
-        "isMinimal",
-        "isWindows",
-        "isLinux",
-        "isMacOS",
-        "isColorSupported",
-        "nodeVersion",
-        "nodeMajorVersion",
-        "process",
-        "providerInfo",
-        "provider",
-        "isNode",
-        "isBun",
-        "isDeno",
-        "isFastly",
-        "isNetlify",
-        "isEdgeLight",
-        "isWorkerd",
-        "isLagon",
-        "runtimeInfo",
-        "runtime",
-      ]
-    `);
-  });
-
   it("defaults", () => {
     expect(stdEnv).toMatchObject({
       platform: expect.any(String),
       provider: expect.any(String),
       isCI: expect.any(Boolean),
-      hasTTY: false,
+      hasTTY: expect.any(Boolean),
       hasWindow: false,
       isDebug: false,
       isTest: true,